Added Authentication and Authorization in both frontend and backend part of the application.

+++ routers/posts.py
@@ -12,6 +12,7 @@ from schemas import (
     PostCreate,
     PostUpdate
 )
+from auth import CurrentUser
 
 router = APIRouter()
 
@@ -30,21 +31,16 @@ async def get_posts(db: Annotated[AsyncSession, Depends(get_db)]):
         response_model=PostResponse,
         status_code=status.HTTP_201_CREATED,
 )
-async def create_post(post: PostCreate, db:Annotated[AsyncSession, Depends(get_db)]):
-    result = await db.execute(
-        select(models.User)
-        .where(models.User.id == post.user_id)
-    )
-    user = result.scalar()
-    if not user:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="User Not Found",
-        )
+async def create_post(
+        post: PostCreate,
+        current_user: CurrentUser,
+        db:Annotated[AsyncSession, Depends(get_db)]
+):
+    
     new_post = models.Post(
         title=post.title,
         content=post.content,
-        user_id=post.user_id,
+        user_id=current_user.id,
     )
     db.add(new_post)
     await db.commit()
@@ -69,6 +65,7 @@ async def get_post(post_id: int, db: Annotated[AsyncSession, Depends(get_db)]):
 async def update_post_full(
     post_id: int,
     post_data: PostCreate, 
+    current_user: CurrentUser,
     db: Annotated[AsyncSession, Depends(get_db)]
 ):
     result = await db.execute(
@@ -84,20 +81,14 @@ async def update_post_full(
             detail="Post Not Found"
         )
 
-    if post_data.user_id != post.user_id:
-        result = await db.execute(
-            select(models.User).where(models.User.id == post_data.user_id)
+    if post.user_id != current_user.id:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="You are not authorized to update this post."
         )
-        user = result.scalar()
-        if not user:
-            raise HTTPException(
-                status_code=status.HTTP_404_NOT_FOUND,
-                detail="User Not Found",
-            )
 
     post.title = post_data.title
     post.content = post_data.content
-    post.user_id = post_data.user_id
 
     await db.commit()
     await db.refresh(post, attribute_names=["author"])
@@ -107,7 +98,8 @@ async def update_post_full(
 @router.patch('/{post_id}', response_model=PostResponse)
 async def update_post_partial(
     post_id: int,
-    post_data: PostUpdate, 
+    post_data: PostUpdate,
+    current_user: CurrentUser,
     db: Annotated[AsyncSession, Depends(get_db)]
 ):
     result = await db.execute(
@@ -116,11 +108,16 @@ async def update_post_partial(
         .where(models.Post.id == post_id)
     )
     post = result.scalar()
-    print(f"Post Scalar() = {post}")
     if not post:
         raise HTTPException(
             status_code=status.HTTP_404_NOT_FOUND,
             detail="Post Not Found"
+        )
+    
+    if post.user_id != current_user.id:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="You are not authorized to update this post."
         )
 
     update_data = post_data.model_dump(exclude_unset=True)
@@ -134,7 +131,11 @@ async def update_post_partial(
     return post
 
 @router.delete('/{post_id}', status_code=status.HTTP_204_NO_CONTENT)
-async def delete_post(post_id: int, db: Annotated[AsyncSession, Depends(get_db)]):
+async def delete_post(
+        post_id: int,
+        current_user: CurrentUser,
+        db: Annotated[AsyncSession, Depends(get_db)]
+):
     result = await db.execute(
         select(models.Post)
         .options(selectinload(models.Post.author))
@@ -145,6 +146,12 @@ async def delete_post(post_id: int, db: Annotated[AsyncSession, Depends(get_db)]
         raise HTTPException(
             status_code=status.HTTP_404_NOT_FOUND,
             detail="Post Not Found"
+        )
+    
+    if post.user_id != current_user.id:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="You are not authorized to delete this post."
         )
 
     await db.delete(post)
